refactor: mobile components folder name
